Create file structure for the project
Issue #38
 - Added Layers to seperate the components of the app
 - Refactored file structure for the project
