+++ app/build.gradle
@@ -59,6 +59,7 @@ dependencies {
     androidTestImplementation 'androidx.test:runner:1.3.0'
     //  version 2 - end
 
+
     //  dependencies used for junit testing    -   end
 
     //  These dependencies are from dev-template branch
