+++ app/src/main/java/redbull/ecard/DataLayer/Result.java
@@ -1,4 +1,4 @@
-package redbull.ecard.data;
+package redbull.ecard.DataLayer;
 
 /**
  * A generic class that holds a result success w/ data or an error exception.
