+++ app/src/main/java/redbull/ecard/UILayer/QuitService.java
@@ -1,10 +1,8 @@
-package redbull.ecard.ui;
+package redbull.ecard.UILayer;
 
 import android.app.Service;
 import android.content.Intent;
 import android.os.IBinder;
-
-import androidx.annotation.Nullable;
 
 import com.google.firebase.auth.FirebaseAuth;
