+++ app/src/main/java/redbull/ecard/UILayer/dashboard/DashboardViewModel.java
@@ -1,4 +1,4 @@
-package redbull.ecard.ui.dashboard;
+package redbull.ecard.UILayer.dashboard;
 
 import androidx.lifecycle.LiveData;
 import androidx.lifecycle.MutableLiveData;
