+++ app/src/main/java/redbull/ecard/UILayer/login/LoginFormState.java
@@ -1,4 +1,4 @@
-package redbull.ecard.ui.login;
+package redbull.ecard.UILayer.login;
 
 import androidx.annotation.Nullable;
